# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Volleyball Team Management API",
    "description": "A RESTful API for managing volleyball players and teams with GitHub OAuth authentication.",
    "version": "1.0.0"
  },
  "host": "localhost:7000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/profile": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/profile": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}